clients/macos/app/collectors/screen.py
"""Screen capture collector for macOS."""

import io
from datetime import datetime
from typing import Any, Dict
import structlog

from app.collectors import BaseCollector

logger = structlog.get_logger(__name__)


class ScreenCollector(BaseCollector):
    """Collects screen capture data."""
    
    def __init__(self, api_client, interval: int, quality: int = 80, 
                 max_width: int = 1920, max_height: int = 1080):
        super().__init__(api_client, interval)
        self.quality = quality
        self.max_width = max_width
        self.max_height = max_height
    
    async def initialize(self) -> None:
        """Initialize the screen collector."""
        try:
            # In a real implementation, you would:
            # 1. Check screen recording permissions
            # 2. Initialize screen capture APIs
            # 3. Get display information
            
            logger.info("Screen collector initialized",
                       quality=self.quality,
                       max_resolution=f"{self.max_width}x{self.max_height}")
            
            self._initialized = True
            
        except Exception as e:
            logger.error("Failed to initialize screen collector", error=str(e))
            raise
    
    async def collect(self) -> bool:
        """Collect screen capture data."""
        if not self._initialized:
            logger.warning("Screen collector not initialized")
            return False
        
        try:
            logger.debug("Capturing screen")
            
            # Capture screen
            image_data = await self._capture_screen()
            
            if not image_data:
                logger.warning("No screen data captured")
                return False
            
            metadata = {
                "width": 1920,  # Would get from actual capture
                "height": 1080,  # Would get from actual capture
                "format": "png",
                "quality": self.quality,
                "timestamp": datetime.utcnow().isoformat()
            }
            
            # Send to API
            success = await self.api_client.send_screen_capture(image_data, metadata)
            
            if success:
                self._collection_count += 1
                self._last_collection = datetime.utcnow().isoformat()
                logger.info("Screen capture collected and sent",
                           size_bytes=len(image_data))
            else:
                self._error_count += 1
                logger.error("Failed to send screen capture")
            
            return success
            
        except Exception as e:
            self._error_count += 1
            logger.error("Error collecting screen capture", error=str(e))
            return False
    
    async def _capture_screen(self) -> bytes:
        """Capture the screen (real implementation would use PyObjC)."""
        # This would be implemented using PyObjC and Core Graphics:
        #
        # from Quartz import CGDisplayCreateImage, CGImageDestinationCreateWithData
        # from CoreFoundation import CFDataCreateMutable
        # 
        # # Get main display
        # display_id = CGMainDisplayID()
        # 
        # # Capture screen
        # image = CGDisplayCreateImage(display_id)
        # 
        # # Convert to PNG data
        # data = CFDataCreateMutable(None, 0)
        # dest = CGImageDestinationCreateWithData(data, "public.png", 1, None)
        # CGImageDestinationAddImage(dest, image, None)
        # CGImageDestinationFinalize(dest)
        # 
        # return bytes(data)
        
        # For now, return mock PNG data
        return self._generate_mock_image_data()
    
    def _generate_mock_image_data(self) -> bytes:
        """Generate mock PNG image data."""
        try:
            from PIL import Image, ImageDraw
            
            # Create a simple test image
            img = Image.new('RGB', (400, 300), color='lightblue')
            draw = ImageDraw.Draw(img)
            
            # Draw some content
            draw.rectangle([50, 50, 350, 250], outline='black', fill='white')
            draw.text((100, 150), "Mock Screen Capture", fill='black')
            
            # Convert to PNG bytes
            buffer = io.BytesIO()
            img.save(buffer, format='PNG', quality=self.quality)
            return buffer.getvalue()
            
        except ImportError:
            # Fallback if PIL not available
            # Return minimal PNG header
            return b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x01\x90\x00\x00\x04\x38\x08\x02\x00\x00\x00'
    
    async def cleanup(self) -> None:
        """Cleanup screen collector resources."""
        logger.info("Screen collector cleaned up")
    
    async def get_status(self) -> Dict[str, Any]:
        """Get screen collector status."""
        base_status = await super().get_status()
        base_status.update({
            "quality": self.quality,
            "max_width": self.max_width,
            "max_height": self.max_height
        })
        return base_status